Migrate RadioButton tests
